requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3

# === Core Scientific Libraries ===
scip
matplotlib
seaborn
scikit-learn
fastapi
uvicorn[standard]
pandas
numpy
yfinance
pydantic
python-dateutil

# === TensorFlow Stack ===
fastapi 
uvicorn 
motor  
alpaca-py
websocket
websocket-client
tensorflow
tensorflow-estimator
tensorflow-macos
keras
kneed
gekko
shap
tensorboard
tensorboard-data-server
gast
cvxpy
h5py
protobuf
absl-py
opt_einsum
wrapt
six
astunparse
flatbuffers
libclang
termcolor
statsmodels
jupyter
pyPortfolioOpt

# === Qiskit Quantum SDK ===
qiskit==1.4.3
qiskit-aer==0.17.0
qiskit-ibm-runtime==0.24.1
qiskit-algorithms==0.3.1
qiskit-optimization==0.6.1

# === Optional Add-ons (only these are 1.x compatible) ===
qiskit-finance==0.4.0
qiskit-optimization==0.6.1

# === Optimization, Modeling, Forecasting ===
docplex==2.29.245
fastdtw==0.3.4
peewee==3.18.1
psutil==7.0.0

# === Finance & Market Data ===
yfinance==0.2.58
pandas-datareader==0.10.0
Nasdaq-Data-Link==1.0.4

# === IBM Cloud SDKs ===
ibm-cloud-sdk-core
ibm-platform-services


# === Data Parsing / Visualization ===
beautifulsoup4==4.13.4
soupsieve==2.7
lxml==5.4.0
Markdown==3.7
MarkupSafe==2.1.5
pyparsing==3.1.4
cycler==0.12.1
kiwisolver==1.4.7
fonttools==4.57.0
pillow==10.4.0
Werkzeug==3.0.6

# === Misc. Packages ===
curl_cffi==0.9.0
dill==0.4.0
more-itertools==10.5.0
frozendict==2.4.6
inflection==0.5.1
multitasking==0.0.11
networkx==3.1
cachetools==5.5.2
websocket-client==1.8.0
